requirements.txt
click==7.1.2
docxtpl==0.10.0
Flask==1.1.2
Flask-MySQLdb==0.2.0
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
python-docx==0.8.10
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
dash-bootstrap-components>=0.10.1
docx-mailmerge==0.5.0